[feat][sophia22001]: 폴더 구조 개편
