+++ babel.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  presets: ['module:@react-native/babel-preset', 'nativewind/babel']
+  presets: ['module:@react-native/babel-preset', 'nativewind/babel'],
 };
