+++ src/presentation/scheduleInpInput/components/TeamItem.tsx
