+++ src/presentation/scheduleRegType/component/SelectScheduleBox.tsx
